test(integration): move controller batch 3
